product_return
This is a C# program, developed by C# and mysql. This is a program developed for my friend to enhance the process of returning products to their customers. She works in a small company that help customers to manage social media accounts, so she have to lend products from the customer's shop and then return to their office. After taking some photos they have to send it back to the customers. And usually the packing process take a lot of time since they have to find the product and make sure the product is in the specific order to return them order by order. So I developed this program to help her search through all products and find out which product belongs to which order instead of the traditional way which was searching receipt by receipt to find which product belongs to which order.